message container re-implemented
